[feature] Introduced UploadType enum for better upload type management; refactored multipart upload functions to utilize the new enum;  moved S3 endpoint into a new config folder instead of just putting it in a file for better code management.
