--- src/app/lib/config.ts
@@ -1,4 +1,0 @@
-export const S3_BUCKET_ENDPOINTS = {
-  splat: process.env.AWS_BUCKET_SPLAT_ENDPOINT || null,
-  video: process.env.AWS_BUCKET_VIDEO_ENDPOINT || null,
-};
