+++ src/app/lib/configs/splatUpload.ts
@@ -1,0 +1,5 @@
+export const S3_BUCKET_ENDPOINTS = {
+    splat: process.env.AWS_BUCKET_SPLAT_ENDPOINT || null,
+    video: process.env.AWS_BUCKET_VIDEO_ENDPOINT || null,
+};
+
